Add the retry decorator for all requests
